opea-comps/ExpertQA/requirements.txt
opea-comps
pydantic
fastapi
uvicorn
numpy
PyPDF2
python-dotenv